[SP] update sentence parser. size down some variable.

+++ code-gcc/NMEAparser/NMEAparser-def.h
@@ -138,12 +138,12 @@ typedef struct {
 
 typedef struct {
 	char* pField;
-	unsigned FieldLen;
+	BYTE FieldLen;
 } NMEA_sentence_field;
 
 
 typedef struct {
-	unsigned vaildfields;
+	BYTE vaildfields;
 	NMEA_sentence_field Field[SENTENCE_FIELD_MAX];
 } NMEA_sentence_field_list;
 
@@ -152,7 +152,7 @@ typedef int (*nmea_decoder) ( unsigned count, NMEA_sentence_field_list* psentenc
 
 typedef struct {
 	char *NMEAID;
-	unsigned fieldNum;			/* minimum number of fields required to parse */
+	BYTE fieldNum;			/* minimum number of fields required to parse */
 	nmea_decoder decoder;
 } NMEA_sentence_decode_list;
 
@@ -160,7 +160,7 @@ typedef int (*Handle_sentence_fields) ( NMEA_sentence_field* psentenceField, gps
 
 typedef struct {
 	BYTE num;
-	BYTE valid;			/* minimum number of fields required to parse */
+	BYTE valid;			/* field handle assigned */
 	Handle_sentence_fields fields_handle;
 } NMEA_sentence_handle_list;
